requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
openai>=1.40.0
requests>=2.32.3
